+++ sqllin-dsl-test/src/jvmTest/kotlin/com/ctrip/sqllin/dsl/test/JvmTest.kt
@@ -136,6 +136,15 @@ class JvmTest {
     @Test
     fun testForeignKeyWithoutPragma() = commonTest.testForeignKeyWithoutPragma()
 
+    @Test
+    fun testDefaultValuesCreateSQL() = commonTest.testDefaultValuesCreateSQL()
+
+    @Test
+    fun testDefaultValuesInsert() = commonTest.testDefaultValuesInsert()
+
+    @Test
+    fun testDefaultValuesWithForeignKey() = commonTest.testDefaultValuesWithForeignKey()
+
     @BeforeTest
     fun setUp() {
         deleteDatabase(path, CommonBasicTest.DATABASE_NAME)